[TD-12505]<feature>: adjustment code
