+++ controller/rest/auth_test.go
