+++ db/tool/createdb.go
@@ -3,7 +3,11 @@ package tool
 import (
 	"unsafe"
 
+	"github.com/taosdata/driver-go/v2/errors"
+	"github.com/taosdata/driver-go/v2/wrapper"
 	"github.com/taosdata/taosadapter/db/async"
+	"github.com/taosdata/taosadapter/httperror"
+	"github.com/taosdata/taosadapter/thread"
 	"github.com/taosdata/taosadapter/tools/pool"
 )
 
@@ -16,6 +20,26 @@ func CreateDBWithConnection(connection unsafe.Pointer, db string) error {
 	err := async.GlobalAsync.TaosExecWithoutResult(connection, b.String())
 	if err != nil {
 		return err
+	}
+	return nil
+}
+
+func SelectDB(taosConnect unsafe.Pointer, db string) error {
+	thread.Lock()
+	code := wrapper.TaosSelectDB(taosConnect, db)
+	thread.Unlock()
+	if code != httperror.SUCCESS {
+		if int32(code)&0xffff == errors.TSC_DB_NOT_SELECTED || int32(code)&0xffff == errors.MND_INVALID_DB {
+			err := CreateDBWithConnection(taosConnect, db)
+			if err != nil {
+				return err
+			}
+			thread.Lock()
+			wrapper.TaosSelectDB(taosConnect, db)
+			thread.Unlock()
+		} else {
+			return errors.GetError(code)
+		}
 	}
 	return nil
 }
