+++ go.mod
@@ -12,12 +12,12 @@ require (
 	github.com/gin-contrib/gzip v0.0.3
 	github.com/gin-contrib/pprof v1.3.0
 	github.com/gin-gonic/gin v1.7.4
-	github.com/gofrs/uuid v4.0.0+incompatible // indirect
 	github.com/influxdata/influxdb/v2 v2.0.9
 	github.com/influxdata/telegraf v1.20.0
 	github.com/lestrrat-go/file-rotatelogs v2.4.0+incompatible
 	github.com/lestrrat-go/strftime v1.0.5 // indirect
 	github.com/patrickmn/go-cache v2.1.0+incompatible
+	github.com/prometheus/prometheus v1.8.2-0.20211209112917-bd21aafb66d6 // indirect
 	github.com/shopspring/decimal v1.2.0 // indirect
 	github.com/silenceper/pool v1.0.0
 	github.com/sirupsen/logrus v1.8.1
@@ -27,6 +27,4 @@ require (
 	github.com/swaggo/swag v1.7.6
 	github.com/taosdata/driver-go/v2 v2.0.1-0.20211215031937-7da3cc9e4ad1
 	github.com/valyala/fastjson v1.6.3
-	go.uber.org/automaxprocs v1.4.0
-	golang.org/x/net v0.0.0-20210805182204-aaa1db679c0d // indirect
 )
