+++ schemaless/capi/opentsdb.go
@@ -6,6 +6,7 @@ import (
 
 	tErrors "github.com/taosdata/driver-go/v2/errors"
 	"github.com/taosdata/driver-go/v2/wrapper"
+	"github.com/taosdata/taosadapter/db/tool"
 	"github.com/taosdata/taosadapter/thread"
 )
 
@@ -13,7 +14,7 @@ func InsertOpentsdbJson(taosConnect unsafe.Pointer, data []byte, db string) erro
 	if len(data) == 0 {
 		return nil
 	}
-	err := SelectDB(taosConnect, db)
+	err := tool.SelectDB(taosConnect, db)
 	if err != nil {
 		return err
 	}
@@ -38,7 +39,7 @@ func InsertOpentsdbTelnet(taosConnect unsafe.Pointer, data, db string) error {
 	if len(data) == 0 {
 		return nil
 	}
-	err := SelectDB(taosConnect, db)
+	err := tool.SelectDB(taosConnect, db)
 	if err != nil {
 		return err
 	}
