+++ system/controller.go
@@ -1,0 +1,3 @@
+package system
+
+import _ "github.com/taosdata/taosadapter/controller/rest"
